Changing return save product

+++ src/main/java/com/example/apitrocatinesql/repositories/TagRepository.java
@@ -4,7 +4,7 @@ import com.example.apitrocatinesql.models.Product;
 import com.example.apitrocatinesql.models.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TagRepository extends JpaRepository<Tag, Integer> {
+public interface TagRepository extends JpaRepository<Tag, Long> {
 
     public Tag findTagByNameIgnoreCase(String name);
